chore: merge multidut pfcwd folder
