+++ tests/snappi_tests/pfcwd/files/pfcwd_basic_helper.py
@@ -3,26 +3,27 @@ from math import ceil
 import logging
 
 from tests.common.helpers.assertions import pytest_assert
-from tests.common.fixtures.conn_graph_facts import conn_graph_facts,\
-    fanout_graph_facts          # noqa F401
-from tests.common.snappi_tests.snappi_helpers import get_dut_port_id
-from tests.common.snappi_tests.common_helpers import pfc_class_enable_vector,\
-    get_pfcwd_poll_interval, get_pfcwd_detect_time, get_pfcwd_restore_time,\
-    enable_packet_aging, start_pfcwd, sec_to_nanosec
-from tests.common.snappi_tests.port import select_ports, select_tx_port
-from tests.common.snappi_tests.snappi_helpers import wait_for_arp
+from tests.common.fixtures.conn_graph_facts import conn_graph_facts, fanout_graph_facts     # noqa: F401
+from tests.common.snappi_tests.snappi_helpers import get_dut_port_id                              # noqa: F401
+from tests.common.snappi_tests.common_helpers import pfc_class_enable_vector, \
+    get_pfcwd_poll_interval, get_pfcwd_detect_time, get_pfcwd_restore_time, \
+    enable_packet_aging, start_pfcwd, sec_to_nanosec, get_pfcwd_stats                             # noqa: F401
+from tests.common.snappi_tests.port import select_ports, select_tx_port                           # noqa: F401
+from tests.common.snappi_tests.snappi_helpers import wait_for_arp                                 # noqa: F401
+from tests.common.snappi_tests.snappi_test_params import SnappiTestParams
 from tests.common.snappi_tests.variables import pfcQueueGroupSize, pfcQueueValueDict
 
 logger = logging.getLogger(__name__)
 
 PAUSE_FLOW_NAME = "Pause Storm"
-WARM_UP_TRAFFIC_NAME = "Warm Up Traffic"
 DATA_FLOW1_NAME = "Data Flow 1"
 DATA_FLOW2_NAME = "Data Flow 2"
+WARM_UP_TRAFFIC_NAME = "Warm Up Traffic"
 WARM_UP_TRAFFIC_DUR = 1
 DATA_PKT_SIZE = 1024
 SNAPPI_POLL_DELAY_SEC = 2
 DEVIATION = 0.3
+UDP_PORT_START = 5000
 
 
 def run_pfcwd_basic_test(api,
@@ -30,11 +31,11 @@ def run_pfcwd_basic_test(api,
                          port_config_list,
                          conn_data,
                          fanout_data,
-                         duthost,
                          dut_port,
                          prio_list,
                          prio_dscp_map,
-                         trigger_pfcwd):
+                         trigger_pfcwd,
+                         snappi_extra_params=None):
     """
     Run a basic PFC watchdog test
 
@@ -49,34 +50,48 @@ def run_pfcwd_basic_test(api,
         prio_list (list): priorities of data flows and pause storm
         prio_dscp_map (dict): Priority vs. DSCP map (key = priority).
         trigger_pfcwd (bool): if PFC watchdog is expected to be triggered
+        snappi_extra_params (SnappiTestParams obj): additional parameters for Snappi traffic
 
     Returns:
         N/A
     """
+    if snappi_extra_params is None:
+        snappi_extra_params = SnappiTestParams()
 
-    pytest_assert(testbed_config is not None,
-                  'Fail to get L2/3 testbed config')
+    # Traffic flow:
+    # tx_port (TGEN) --- ingress DUT --- egress DUT --- rx_port (TGEN)
 
-    start_pfcwd(duthost)
-    enable_packet_aging(duthost)
+    rx_port = snappi_extra_params.multi_dut_params.multi_dut_ports[0]
+    egress_duthost = rx_port['duthost']
+
+    tx_port = snappi_extra_params.multi_dut_params.multi_dut_ports[1]
+    ingress_duthost = tx_port["duthost"]
+    pytest_assert(testbed_config is not None, 'Fail to get L2/3 testbed config')
+
+    if (egress_duthost.is_multi_asic):
+        enable_packet_aging(egress_duthost, rx_port['asic_value'])
+        enable_packet_aging(ingress_duthost, tx_port['asic_value'])
+        start_pfcwd(egress_duthost, rx_port['asic_value'])
+        start_pfcwd(ingress_duthost, tx_port['asic_value'])
+    else:
+        enable_packet_aging(egress_duthost)
+        enable_packet_aging(ingress_duthost)
+        start_pfcwd(egress_duthost)
+        start_pfcwd(ingress_duthost)
+
+    ini_stats = {}
+    for prio in prio_list:
+        ini_stats.update(get_stats(egress_duthost, rx_port['peer_port'], prio))
 
     # Set appropriate pfcwd loss deviation - these values are based on empirical testing
-    DEVIATION = 0.35 if duthost.facts['asic_type'] in ["broadcom"] else 0.3
+    DEVIATION = 0.35 if egress_duthost.facts['asic_type'] in ["broadcom"] or \
+        ingress_duthost.facts['asic_type'] in ["broadcom"] else 0.3
 
-    """ Get the ID of the port to test """
-    port_id = get_dut_port_id(dut_hostname=duthost.hostname,
-                              dut_port=dut_port,
-                              conn_data=conn_data,
-                              fanout_data=fanout_data)
-
-    pytest_assert(port_id is not None,
-                  'Fail to get ID for port {}'.format(dut_port))
-
-    poll_interval_sec = get_pfcwd_poll_interval(duthost) / 1000.0
-    detect_time_sec = get_pfcwd_detect_time(
-        host_ans=duthost, intf=dut_port) / 1000.0
-    restore_time_sec = get_pfcwd_restore_time(
-        host_ans=duthost, intf=dut_port) / 1000.0
+    poll_interval_sec = get_pfcwd_poll_interval(egress_duthost, rx_port['asic_value']) / 1000.0
+    detect_time_sec = get_pfcwd_detect_time(host_ans=egress_duthost, intf=dut_port,
+                                            asic_value=rx_port['asic_value']) / 1000.0
+    restore_time_sec = get_pfcwd_restore_time(host_ans=egress_duthost, intf=dut_port,
+                                              asic_value=rx_port['asic_value']) / 1000.0
 
     """ Warm up traffic is initially sent before any other traffic to prevent pfcwd
     fake alerts caused by idle links (non-incremented packet counters) during pfcwd detection periods """
@@ -111,11 +126,12 @@ def run_pfcwd_basic_test(api,
         flow1_min_loss_rate = 0
 
     exp_dur_sec = flow2_delay_sec + flow2_dur_sec + 1
+    cisco_platform = "Cisco" in egress_duthost.facts['hwsku']
 
     """ Generate traffic config """
     __gen_traffic(testbed_config=testbed_config,
                   port_config_list=port_config_list,
-                  port_id=port_id,
+                  port_id=0,
                   pause_flow_name=PAUSE_FLOW_NAME,
                   pause_flow_dur_sec=pfc_storm_dur_sec,
                   data_flow_name_list=[WARM_UP_TRAFFIC_NAME,
@@ -126,7 +142,9 @@ def run_pfcwd_basic_test(api,
                       warm_up_traffic_dur_sec, flow1_dur_sec, flow2_dur_sec],
                   data_pkt_size=DATA_PKT_SIZE,
                   prio_list=prio_list,
-                  prio_dscp_map=prio_dscp_map)
+                  prio_dscp_map=prio_dscp_map,
+                  traffic_rate=49.99 if cisco_platform else 100.0,
+                  number_of_streams=1)
 
     flows = testbed_config.flows
 
@@ -137,10 +155,47 @@ def run_pfcwd_basic_test(api,
                                all_flow_names=all_flow_names,
                                exp_dur_sec=exp_dur_sec)
 
+    fin_stats = {}
+    for prio in prio_list:
+        fin_stats.update(get_stats(egress_duthost, rx_port['peer_port'], prio))
+
+    loss_packets = 0
+    for k in fin_stats.keys():
+        logger.info('Parameter:{}, Initial Value:{}, Final Value:{}'.format(k, ini_stats[k], fin_stats[k]))
+        if 'DROP' in k:
+            loss_packets += (int(fin_stats[k]) - int(ini_stats[k]))
+
+    logger.info('Total PFCWD drop packets before and after the test:{}'.format(loss_packets))
+
     __verify_results(rows=flow_stats,
                      data_flow_name_list=[DATA_FLOW1_NAME, DATA_FLOW2_NAME],
                      data_flow_min_loss_rate_list=[flow1_min_loss_rate, 0],
-                     data_flow_max_loss_rate_list=[flow1_max_loss_rate, 0])
+                     data_flow_max_loss_rate_list=[flow1_max_loss_rate, 0],
+                     loss_packets=loss_packets)
+
+
+def get_stats(duthost, port, prio):
+    """
+    Returns the PFCWD stats for Tx Ok, Tx drop, Storm detected and restored.
+
+    Args:
+        duthost (obj): DUT
+        port (string): Port on the DUT
+        prio (int):    Priority
+
+    Returns:
+        Dictionary with prio_'parameter' as key and associated value.
+
+    """
+    my_dict = {}
+    new_dict = {}
+    init_pfcwd = get_pfcwd_stats(duthost, port, prio)
+    key_list = ['TX_OK/DROP', 'STORM_DETECTED/RESTORED']
+    for keys in key_list:
+        my_dict[keys] = init_pfcwd[keys]
+    new_dict = {str(prio)+'_'+k: v for key, value in my_dict.items() for k, v in zip(key.split('/'), value.split('/'))}
+
+    return new_dict
 
 
 def __gen_traffic(testbed_config,
@@ -153,7 +208,9 @@ def __gen_traffic(testbed_config,
                   data_flow_dur_sec_list,
                   data_pkt_size,
                   prio_list,
-                  prio_dscp_map):
+                  prio_dscp_map,
+                  traffic_rate,
+                  number_of_streams):
     """
     Generate configurations of flows, including data flows and pause storm.
 
@@ -169,6 +226,8 @@ def __gen_traffic(testbed_config,
         data_pkt_size (int): size of data packets in byte
         prio_list (list): priorities of data flows and pause storm
         prio_dscp_map (dict): Priority vs. DSCP map (key = priority).
+        traffic_rate: Total rate of traffic for all streams together.
+        number_of_streams: The number of UDP streams needed.
 
     Returns:
         N/A
@@ -241,7 +300,7 @@ def __gen_traffic(testbed_config,
 
     tx_port_name = testbed_config.ports[tx_port_id].name
     rx_port_name = testbed_config.ports[rx_port_id].name
-    data_flow_rate_percent = int(100 / len(prio_list))
+    data_flow_rate_percent = int(traffic_rate / len(prio_list))
 
     """ For each data flow """
     for i in range(len(data_flow_name_list)):
@@ -253,13 +312,21 @@ def __gen_traffic(testbed_config,
             data_flow.tx_rx.port.tx_name = tx_port_name
             data_flow.tx_rx.port.rx_name = rx_port_name
 
-            eth, ipv4 = data_flow.packet.ethernet().ipv4()
+            eth, ipv4, udp = data_flow.packet.ethernet().ipv4().udp()
+
             eth.src.value = tx_mac
             eth.dst.value = rx_mac
             if pfcQueueGroupSize == 8:
                 eth.pfc_queue.value = prio
             else:
                 eth.pfc_queue.value = pfcQueueValueDict[prio]
+
+            global UDP_PORT_START
+            src_port = UDP_PORT_START
+            UDP_PORT_START += number_of_streams
+            udp.src_port.increment.start = src_port
+            udp.src_port.increment.step = 1
+            udp.src_port.increment.count = number_of_streams
 
             ipv4.src.value = tx_port_config.ip
             ipv4.dst.value = rx_port_config.ip
@@ -340,7 +407,8 @@ def __run_traffic(api, config, all_flow_names, exp_dur_sec):
 def __verify_results(rows,
                      data_flow_name_list,
                      data_flow_min_loss_rate_list,
-                     data_flow_max_loss_rate_list):
+                     data_flow_max_loss_rate_list,
+                     loss_packets):
     """
     Verify if we get expected experiment results
 
@@ -367,7 +435,9 @@ def __verify_results(rows,
                 data_flow_tx_frames_list[i] += tx_frames
                 data_flow_rx_frames_list[i] += rx_frames
 
+    tgen_loss_packets = 0
     for i in range(num_data_flows):
+        tgen_loss_packets += data_flow_tx_frames_list[i] - data_flow_rx_frames_list[i]
         loss_rate = 1 - \
             float(data_flow_rx_frames_list[i]) / data_flow_tx_frames_list[i]
         min_loss_rate = data_flow_min_loss_rate_list[i]
@@ -376,3 +446,5 @@ def __verify_results(rows,
         pytest_assert(loss_rate <= max_loss_rate and loss_rate >= min_loss_rate,
                       'Loss rate of {} ({}) should be in [{}, {}]'.format(
                           data_flow_name_list[i], loss_rate, min_loss_rate, max_loss_rate))
+
+    logger.info('TGEN Loss packets:{}'.format(tgen_loss_packets))
