+++ tests/snappi_tests/pfcwd/test_pfcwd_basic_with_snappi.py
@@ -1,96 +1,119 @@
-import logging
 import pytest
-
-from tests.common.helpers.assertions import pytest_require, pytest_assert
-from tests.common.fixtures.conn_graph_facts import conn_graph_facts,\
-    fanout_graph_facts                      # noqa F401
-from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port,\
-    snappi_api, snappi_testbed_config       # noqa F401
+import random
+import logging
+import time
+import re
+from collections import defaultdict
+from tests.common.helpers.assertions import pytest_require, pytest_assert                               # noqa: F401
+from tests.common.fixtures.conn_graph_facts import conn_graph_facts, fanout_graph_facts, \
+    fanout_graph_facts_multidut     # noqa: F401
+from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port, \
+    get_snappi_ports_single_dut, snappi_testbed_config, \
+    get_snappi_ports_multi_dut, is_snappi_multidut, \
+    snappi_api, snappi_dut_base_config, get_snappi_ports, get_snappi_ports_for_rdma, cleanup_config      # noqa: F401
 from tests.common.snappi_tests.qos_fixtures import prio_dscp_map, lossless_prio_list      # noqa F401
+from tests.common.reboot import reboot                              # noqa: F401
+from tests.common.utilities import wait_until                       # noqa: F401
 from tests.common.config_reload import config_reload
-from tests.common.reboot import reboot
-from tests.common.utilities import wait_until
+from tests.common.platform.interface_utils import check_interface_status_of_up_ports
 from tests.snappi_tests.pfcwd.files.pfcwd_basic_helper import run_pfcwd_basic_test
-from tests.snappi_tests.files.helper import skip_warm_reboot
-
+from tests.common.snappi_tests.snappi_test_params import SnappiTestParams
+from tests.snappi_tests.files.helper import skip_pfcwd_test, reboot_duts, \
+    setup_ports_and_dut, multidut_port_info   # noqa: F401
 logger = logging.getLogger(__name__)
+pytestmark = [pytest.mark.topology('multidut-tgen', 'tgen')]
 
-pytestmark = [pytest.mark.topology('tgen')]
+WAIT_TIME = 600
+INTERVAL = 40
+
+
+@pytest.fixture(autouse=True)
+def number_of_tx_rx_ports():
+    yield (1, 1)
+
+
+@pytest.fixture(autouse=False)
+def save_restore_config(setup_ports_and_dut):          # noqa: F811
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
+    timestamp = time.time()
+    dest = f'~/{timestamp}'
+
+    for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+        duthost.shell(f"sudo mkdir {dest}; sudo cp /etc/sonic/config*.json {dest}")
+        duthost.shell("sudo config save -y")
+
+    yield
+
+    for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+        duthost.shell(f"sudo cp {dest}/config_db*json /etc/sonic/")
+
+    for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+        config_reload(duthost)
 
 
 @pytest.mark.parametrize("trigger_pfcwd", [True, False])
-def test_pfcwd_basic_single_lossless_prio(snappi_api,               # noqa F811
-                                          snappi_testbed_config,    # noqa F811
-                                          conn_graph_facts,         # noqa F811
-                                          fanout_graph_facts,       # noqa F811
+def test_pfcwd_basic_single_lossless_prio(snappi_api,                   # noqa: F811
+                                          conn_graph_facts,             # noqa: F811
+                                          fanout_graph_facts_multidut,           # noqa: F811
                                           duthosts,
-                                          rand_one_dut_hostname,
-                                          rand_one_dut_portname_oper_up,
-                                          enum_dut_lossless_prio,
-                                          prio_dscp_map,            # noqa F811
-                                          trigger_pfcwd):
+                                          lossless_prio_list,    # noqa: F811
+                                          tbinfo,                # noqa: F811
+                                          prio_dscp_map,         # noqa F811
+                                          setup_ports_and_dut,   # noqa: F811
+                                          trigger_pfcwd,         # noqa: F811
+                                          ):
     """
     Run PFC watchdog basic test on a single lossless priority
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
-        snappi_testbed_config (pytest fixture): testbed configuration information
         conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts (pytest fixture): fanout graph
+        fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
-        rand_one_dut_hostname (str): hostname of DUT
-        rand_one_dut_portname_oper_up (str): port to test, e.g., 's6100-1|Ethernet0'
-        enum_dut_lossless_prio (str): name of lossless priority to test, e.g., 's6100-1|3'
         prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority)
         trigger_pfcwd (bool): if PFC watchdog is expected to be triggered
 
     Returns:
         N/A
     """
-    dut_hostname, dut_port = rand_one_dut_portname_oper_up.split('|')
-    dut_hostname2, lossless_prio = enum_dut_lossless_prio.split('|')
-    pytest_require(rand_one_dut_hostname == dut_hostname == dut_hostname2,
-                   "Priority and port are not mapped to the expected DUT")
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
 
-    duthost = duthosts[rand_one_dut_hostname]
+    lossless_prio = random.sample(lossless_prio_list, 1)
+    lossless_prio = int(lossless_prio[0])
 
-    testbed_config, port_config_list = snappi_testbed_config
-    lossless_prio = int(lossless_prio)
+    snappi_extra_params = SnappiTestParams()
+    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
 
     run_pfcwd_basic_test(api=snappi_api,
                          testbed_config=testbed_config,
                          port_config_list=port_config_list,
                          conn_data=conn_graph_facts,
-                         fanout_data=fanout_graph_facts,
-                         duthost=duthost,
-                         dut_port=dut_port,
+                         fanout_data=fanout_graph_facts_multidut,
+                         dut_port=snappi_ports[0]['peer_port'],
                          prio_list=[lossless_prio],
                          prio_dscp_map=prio_dscp_map,
-                         trigger_pfcwd=trigger_pfcwd)
+                         trigger_pfcwd=trigger_pfcwd,
+                         snappi_extra_params=snappi_extra_params)
 
 
 @pytest.mark.parametrize("trigger_pfcwd", [True, False])
 def test_pfcwd_basic_multi_lossless_prio(snappi_api,                # noqa F811
-                                         snappi_testbed_config,     # noqa F811
                                          conn_graph_facts,          # noqa F811
-                                         fanout_graph_facts,        # noqa F811
+                                         fanout_graph_facts_multidut,        # noqa F811
                                          duthosts,
-                                         rand_one_dut_hostname,
-                                         rand_one_dut_portname_oper_up,
-                                         lossless_prio_list,        # noqa F811
+                                         lossless_prio_list,    # noqa: F811
+                                         tbinfo,      # noqa: F811
                                          prio_dscp_map,             # noqa F811
+                                         setup_ports_and_dut,       # noqa: F811
                                          trigger_pfcwd):
     """
     Run PFC watchdog basic test on multiple lossless priorities
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
-        snappi_testbed_config (pytest fixture): testbed configuration information
         conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts (pytest fixture): fanout graph
+        fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
-        rand_one_dut_hostname (str): hostname of DUT
-        rand_one_dut_portname_oper_up (str): port to test, e.g., 's6100-1|Ethernet0'
         lossless_prio_list (pytest fixture): list of all the lossless priorities
         prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority)
         trigger_pfcwd (bool): if PFC watchdog is expected to be triggered
@@ -98,178 +121,138 @@ def test_pfcwd_basic_multi_lossless_prio(snappi_api,                # noqa F811
     Returns:
         N/A
     """
-    dut_hostname, dut_port = rand_one_dut_portname_oper_up.split('|')
-    pytest_require(rand_one_dut_hostname == dut_hostname,
-                   "Port is not mapped to the expected DUT")
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
 
-    duthost = duthosts[rand_one_dut_hostname]
-
-    testbed_config, port_config_list = snappi_testbed_config
+    snappi_extra_params = SnappiTestParams()
+    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
 
     run_pfcwd_basic_test(api=snappi_api,
                          testbed_config=testbed_config,
                          port_config_list=port_config_list,
                          conn_data=conn_graph_facts,
-                         fanout_data=fanout_graph_facts,
-                         duthost=duthost,
-                         dut_port=dut_port,
+                         fanout_data=fanout_graph_facts_multidut,
+                         dut_port=snappi_ports[0]['peer_port'],
                          prio_list=lossless_prio_list,
                          prio_dscp_map=prio_dscp_map,
-                         trigger_pfcwd=trigger_pfcwd)
+                         trigger_pfcwd=trigger_pfcwd,
+                         snappi_extra_params=snappi_extra_params)
 
 
 @pytest.mark.disable_loganalyzer
-@pytest.mark.parametrize('reboot_type', ['warm', 'cold', 'fast'])
 @pytest.mark.parametrize("trigger_pfcwd", [True, False])
 def test_pfcwd_basic_single_lossless_prio_reboot(snappi_api,                # noqa F811
-                                                 snappi_testbed_config,     # noqa F811
                                                  conn_graph_facts,          # noqa F811
-                                                 fanout_graph_facts,        # noqa F811
+                                                 fanout_graph_facts_multidut,        # noqa F811
                                                  localhost,
                                                  duthosts,
-                                                 rand_one_dut_hostname,
-                                                 rand_one_dut_portname_oper_up,
-                                                 rand_one_dut_lossless_prio,
+                                                 enum_dut_lossless_prio_with_completeness_level,   # noqa: F811
+                                                 get_snappi_ports,   # noqa: F811
                                                  prio_dscp_map,             # noqa F811
-                                                 reboot_type,
+                                                 setup_ports_and_dut,    # noqa: F811
+                                                 reboot_duts,               # noqa: F811
                                                  trigger_pfcwd):
     """
     Verify PFC watchdog basic test works on a single lossless priority after various types of reboot
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
-        snappi_testbed_config (pytest fixture): testbed configuration information
         conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts (pytest fixture): fanout graph
+        fanout_graph_facts_multidut (pytest fixture): fanout graph
         localhost (pytest fixture): localhost handle
         duthosts (pytest fixture): list of DUTs
-        rand_one_dut_hostname (str): hostname of DUT
-        rand_one_dut_portname_oper_up (str): name of port to test, e.g., 's6100-1|Ethernet0'
-        rand_one_dut_lossless_prio (str): name of lossless priority to test, e.g., 's6100-1|3'
         prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority)
-        reboot_type (str): reboot type to be issued on the DUT
         trigger_pfcwd (bool): if PFC watchdog is expected to be triggered
 
     Returns:
         N/A
     """
-    dut_hostname, dut_port = rand_one_dut_portname_oper_up.split('|')
-    dut_hostname2, lossless_prio = rand_one_dut_lossless_prio.split('|')
-    pytest_require(rand_one_dut_hostname == dut_hostname == dut_hostname2,
-                   "Priority and port are not mapped to the expected DUT")
 
-    duthost = duthosts[rand_one_dut_hostname]
-    skip_warm_reboot(duthost, reboot_type)
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
 
-    testbed_config, port_config_list = snappi_testbed_config
+    _, lossless_prio = enum_dut_lossless_prio_with_completeness_level.split('|')
     lossless_prio = int(lossless_prio)
-
-    logger.info("Issuing a {} reboot on the dut {}".format(reboot_type, duthost.hostname))
-    reboot(duthost, localhost, reboot_type=reboot_type, safe_reboot=True)
-    logger.info("Wait until the system is stable")
-    pytest_assert(wait_until(300, 20, 0, duthost.critical_services_fully_started),
-                  "Not all critical services are fully started")
+    snappi_extra_params = SnappiTestParams()
+    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
 
     run_pfcwd_basic_test(api=snappi_api,
                          testbed_config=testbed_config,
                          port_config_list=port_config_list,
                          conn_data=conn_graph_facts,
-                         fanout_data=fanout_graph_facts,
-                         duthost=duthost,
-                         dut_port=dut_port,
+                         fanout_data=fanout_graph_facts_multidut,
+                         dut_port=snappi_ports[0]['peer_port'],
                          prio_list=[lossless_prio],
                          prio_dscp_map=prio_dscp_map,
-                         trigger_pfcwd=trigger_pfcwd)
+                         trigger_pfcwd=trigger_pfcwd,
+                         snappi_extra_params=snappi_extra_params)
 
 
 @pytest.mark.disable_loganalyzer
-@pytest.mark.parametrize('reboot_type', ['warm', 'cold', 'fast'])
 @pytest.mark.parametrize("trigger_pfcwd", [True, False])
 def test_pfcwd_basic_multi_lossless_prio_reboot(snappi_api,                 # noqa F811
-                                                snappi_testbed_config,      # noqa F811
                                                 conn_graph_facts,           # noqa F811
-                                                fanout_graph_facts,         # noqa F811
+                                                fanout_graph_facts_multidut,         # noqa F811
                                                 localhost,
-                                                duthosts,
-                                                rand_one_dut_hostname,
-                                                rand_one_dut_portname_oper_up,
                                                 lossless_prio_list,         # noqa F811
+                                                tbinfo,      # noqa: F811
                                                 prio_dscp_map,              # noqa F811
-                                                reboot_type,
+                                                setup_ports_and_dut,        # noqa: F811
+                                                reboot_duts,                # noqa: F811
                                                 trigger_pfcwd):
     """
     Verify PFC watchdog basic test works on multiple lossless priorities after various kinds of reboots
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
-        snappi_testbed_config (pytest fixture): testbed configuration information
         conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts (pytest fixture): fanout graph
+        fanout_graph_facts_multidut (pytest fixture): fanout graph
         localhost (pytest fixture): localhost handle
         duthosts (pytest fixture): list of DUTs
-        rand_one_dut_hostname (str): hostname of DUT
-        rand_one_dut_portname_oper_up (str): name of port to test, e.g., 's6100-1|Ethernet0'
         lossless_prio_list (pytest fixture): list of all the lossless priorities
         prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority)
-        reboot_type (str): reboot type to be issued on the DUT
         trigger_pfcwd (bool): if PFC watchdog is expected to be triggered
 
     Returns:
         N/A
     """
-    dut_hostname, dut_port = rand_one_dut_portname_oper_up.split('|')
-    pytest_require(rand_one_dut_hostname == dut_hostname,
-                   "Port is not mapped to the expected DUT")
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
 
-    duthost = duthosts[rand_one_dut_hostname]
-    skip_warm_reboot(duthost, reboot_type)
-
-    testbed_config, port_config_list = snappi_testbed_config
-
-    logger.info("Issuing a {} reboot on the dut {}".format(reboot_type, duthost.hostname))
-    reboot(duthost, localhost, reboot_type=reboot_type, safe_reboot=True)
-    logger.info("Wait until the system is stable")
-    pytest_assert(wait_until(300, 20, 0, duthost.critical_services_fully_started),
-                  "Not all critical services are fully started")
+    snappi_extra_params = SnappiTestParams()
+    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
 
     run_pfcwd_basic_test(api=snappi_api,
                          testbed_config=testbed_config,
                          port_config_list=port_config_list,
                          conn_data=conn_graph_facts,
-                         fanout_data=fanout_graph_facts,
-                         duthost=duthost,
-                         dut_port=dut_port,
+                         fanout_data=fanout_graph_facts_multidut,
+                         dut_port=snappi_ports[0]['peer_port'],
                          prio_list=lossless_prio_list,
                          prio_dscp_map=prio_dscp_map,
-                         trigger_pfcwd=trigger_pfcwd)
+                         trigger_pfcwd=trigger_pfcwd,
+                         snappi_extra_params=snappi_extra_params)
 
 
 @pytest.mark.disable_loganalyzer
 @pytest.mark.parametrize('restart_service', ['swss'])
 @pytest.mark.parametrize("trigger_pfcwd", [True, False])
 def test_pfcwd_basic_single_lossless_prio_service_restart(snappi_api,               # noqa F811
-                                                          snappi_testbed_config,    # noqa F811
                                                           conn_graph_facts,         # noqa F811
-                                                          fanout_graph_facts,       # noqa F811
+                                                          fanout_graph_facts_multidut,       # noqa F811
                                                           duthosts,
-                                                          rand_one_dut_hostname,
-                                                          rand_one_dut_portname_oper_up,
-                                                          rand_one_dut_lossless_prio,
-                                                          prio_dscp_map,            # noqa F811
+                                                          lossless_prio_list,   # noqa: F811
+                                                          tbinfo,      # noqa: F811
+                                                          prio_dscp_map,            # noqa: F811
                                                           restart_service,
-                                                          trigger_pfcwd):
+                                                          trigger_pfcwd,
+                                                          setup_ports_and_dut,      # noqa: F811
+                                                          save_restore_config):
     """
     Verify PFC watchdog basic test works on a single lossless priority after various service restarts
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
-        snappi_testbed_config (pytest fixture): testbed configuration information
         conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts (pytest fixture): fanout graph
+        fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
-        rand_one_dut_hostname (str): hostname of DUT
-        rand_one_dut_portname_oper_up (str): name of port to test, e.g., 's6100-1|Ethernet0'
-        rand_one_dut_lossless_prio (str): name of lossless priority to test, e.g., 's6100-1|3'
         prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority)
         restart_service (str): service to restart on the DUT. Only 'swss' affects pfcwd currently
         trigger_pfcwd (bool): if PFC watchdog is expected to be triggered
@@ -277,62 +260,85 @@ def test_pfcwd_basic_single_lossless_prio_service_restart(snappi_api,           
     Returns:
         N/A
     """
-    dut_hostname, dut_port = rand_one_dut_portname_oper_up.split('|')
-    dut_hostname2, lossless_prio = rand_one_dut_lossless_prio.split('|')
-    pytest_require(rand_one_dut_hostname == dut_hostname == dut_hostname2,
-                   "Priority and port are not mapped to the expected DUT")
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
+    lossless_prio = random.sample(lossless_prio_list, 1)
+    lossless_prio = int(lossless_prio[0])
 
-    duthost = duthosts[rand_one_dut_hostname]
+    if (snappi_ports[0]['duthost'].is_multi_asic):
+        ports_dict = defaultdict(list)
+        for port in snappi_ports:
+            ports_dict[port['peer_device']].append(port['asic_value'])
 
-    testbed_config, port_config_list = snappi_testbed_config
-    lossless_prio = int(lossless_prio)
+        for k in ports_dict.keys():
+            ports_dict[k] = list(set(ports_dict[k]))
 
-    logger.info("Issuing a restart of service {} on the dut {}".format(restart_service, duthost.hostname))
-    duthost.command("systemctl reset-failed {}".format(restart_service))
-    duthost.command("systemctl restart {}".format(restart_service))
-    logger.info("Wait until the system is stable")
-    pytest_assert(wait_until(300, 20, 0, duthost.critical_services_fully_started),
-                  "Not all critical services are fully started")
+        logger.info('Port dictionary:{}'.format(ports_dict))
+        for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+            up_bgp_neighbors = duthost.get_bgp_neighbors_per_asic("established")
+            # Record current state of critical services.
+            duthost.critical_services_fully_started()
+
+            asic_list = ports_dict[duthost.hostname]
+            asic = random.sample(asic_list, 1)[0]
+            asic_id = re.match(r"(asic)(\d+)", asic).group(2)
+            proc = 'swss@' + asic_id
+            logger.info("Issuing a restart of service {} on the dut {}".format(proc, duthost.hostname))
+            duthost.command("sudo systemctl reset-failed {}".format(proc))
+            duthost.command("sudo systemctl restart {}".format(proc))
+            logger.info("Wait until the system is stable")
+            pytest_assert(wait_until(WAIT_TIME, INTERVAL, 0, duthost.critical_services_fully_started),
+                          "Not all critical services are fully started")
+            pytest_assert(wait_until(WAIT_TIME, INTERVAL, 0, check_interface_status_of_up_ports, duthost),
+                          "Not all interfaces are up.")
+            pytest_assert(wait_until(
+                WAIT_TIME, INTERVAL, 0, duthost.check_bgp_session_state_all_asics, up_bgp_neighbors, "established"))
+
+    else:
+        for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+            logger.info("Issuing a restart of service {} on the dut {}".format(restart_service, duthost.hostname))
+            duthost.command("systemctl reset-failed {}".format(restart_service))
+            duthost.command("systemctl restart {}".format(restart_service))
+            logger.info("Wait until the system is stable")
+            pytest_assert(wait_until(WAIT_TIME, INTERVAL, 0, duthost.critical_services_fully_started),
+                          "Not all critical services are fully started")
+
+    snappi_extra_params = SnappiTestParams()
+    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
 
     run_pfcwd_basic_test(api=snappi_api,
                          testbed_config=testbed_config,
                          port_config_list=port_config_list,
                          conn_data=conn_graph_facts,
-                         fanout_data=fanout_graph_facts,
-                         duthost=duthost,
-                         dut_port=dut_port,
+                         fanout_data=fanout_graph_facts_multidut,
+                         dut_port=snappi_ports[0]['peer_port'],
                          prio_list=[lossless_prio],
                          prio_dscp_map=prio_dscp_map,
-                         trigger_pfcwd=trigger_pfcwd)
-
-    config_reload(sonic_host=duthost, config_source='config_db', safe_reload=True)
+                         trigger_pfcwd=trigger_pfcwd,
+                         snappi_extra_params=snappi_extra_params)
 
 
 @pytest.mark.disable_loganalyzer
 @pytest.mark.parametrize('restart_service', ['swss'])
 @pytest.mark.parametrize("trigger_pfcwd", [True, False])
 def test_pfcwd_basic_multi_lossless_prio_restart_service(snappi_api,                # noqa F811
-                                                         snappi_testbed_config,     # noqa F811
                                                          conn_graph_facts,          # noqa F811
-                                                         fanout_graph_facts,        # noqa F811
+                                                         fanout_graph_facts_multidut,        # noqa F811
                                                          duthosts,
-                                                         rand_one_dut_hostname,
-                                                         rand_one_dut_portname_oper_up,
-                                                         lossless_prio_list,        # noqa F811
+                                                         lossless_prio_list,    # noqa: F811
+                                                         tbinfo,      # noqa: F811
                                                          prio_dscp_map,             # noqa F811
                                                          restart_service,
-                                                         trigger_pfcwd):
+                                                         setup_ports_and_dut,       # noqa: F811
+                                                         trigger_pfcwd,
+                                                         save_restore_config):
     """
     Verify PFC watchdog basic test works on multiple lossless priorities after various service restarts
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
-        snappi_testbed_config (pytest fixture): testbed configuration information
         conn_graph_facts (pytest fixture): connection graph
-        fanout_graph_facts (pytest fixture): fanout graph
+        fanout_graph_facts_multidut (pytest fixture): fanout graph
         duthosts (pytest fixture): list of DUTs
-        rand_one_dut_hostname (str): hostname of DUT
-        rand_one_dut_portname_oper_up (str): name of port to test, e.g., 's6100-1|Ethernet0'
         lossless_prio_list (pytest fixture): list of all the lossless priorities
         prio_dscp_map (pytest fixture): priority vs. DSCP map (key = priority)
         restart_service (str): service to restart on the DUT. Only 'swss' affects pfcwd currently
@@ -341,30 +347,57 @@ def test_pfcwd_basic_multi_lossless_prio_restart_service(snappi_api,            
     Returns:
         N/A
     """
-    dut_hostname, dut_port = rand_one_dut_portname_oper_up.split('|')
-    pytest_require(rand_one_dut_hostname == dut_hostname,
-                   "Port is not mapped to the expected DUT")
 
-    duthost = duthosts[rand_one_dut_hostname]
+    testbed_config, port_config_list, snappi_ports = setup_ports_and_dut
 
-    testbed_config, port_config_list = snappi_testbed_config
+    if (snappi_ports[0]['duthost'].is_multi_asic):
+        ports_dict = defaultdict(list)
+        for port in snappi_ports:
+            ports_dict[port['peer_device']].append(port['asic_value'])
 
-    logger.info("Issuing a restart of service {} on the dut {}".format(restart_service, duthost.hostname))
-    duthost.command("systemctl reset-failed {}".format(restart_service))
-    duthost.command("systemctl restart {}".format(restart_service))
-    logger.info("Wait until the system is stable")
-    pytest_assert(wait_until(300, 20, 0, duthost.critical_services_fully_started),
-                  "Not all critical services are fully started")
+        for k in ports_dict.keys():
+            ports_dict[k] = list(set(ports_dict[k]))
 
+        logger.info('Port dictionary:{}'.format(ports_dict))
+        for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+            up_bgp_neighbors = duthost.get_bgp_neighbors_per_asic("established")
+            # Record current state of critical services.
+            duthost.critical_services_fully_started()
+
+            asic_list = ports_dict[duthost.hostname]
+            asic = random.sample(asic_list, 1)[0]
+            asic_id = re.match(r"(asic)(\d+)", asic).group(2)
+            proc = 'swss@' + asic_id
+
+            logger.info("Issuing a restart of service {} on the dut {}".format(proc, duthost.hostname))
+            duthost.command("sudo systemctl reset-failed {}".format(proc))
+            duthost.command("sudo systemctl restart {}".format(proc))
+            logger.info("Wait until the system is stable")
+            pytest_assert(wait_until(WAIT_TIME, INTERVAL, 0, duthost.critical_services_fully_started),
+                          "Not all critical services are fully started")
+            pytest_assert(wait_until(WAIT_TIME, INTERVAL, 0, check_interface_status_of_up_ports, duthost),
+                          "Not all interfaces are up.")
+            pytest_assert(wait_until(
+                WAIT_TIME, INTERVAL, 0, duthost.check_bgp_session_state_all_asics, up_bgp_neighbors, "established"))
+
+    else:
+        for duthost in list(set([snappi_ports[0]['duthost'], snappi_ports[1]['duthost']])):
+            logger.info("Issuing a restart of service {} on the dut {}".format(restart_service, duthost.hostname))
+            duthost.command("systemctl reset-failed {}".format(restart_service))
+            duthost.command("systemctl restart {}".format(restart_service))
+            logger.info("Wait until the system is stable")
+            pytest_assert(wait_until(WAIT_TIME, INTERVAL, 0, duthost.critical_services_fully_started),
+                          "Not all critical services are fully started")
+
+    snappi_extra_params = SnappiTestParams()
+    snappi_extra_params.multi_dut_params.multi_dut_ports = snappi_ports
     run_pfcwd_basic_test(api=snappi_api,
                          testbed_config=testbed_config,
                          port_config_list=port_config_list,
                          conn_data=conn_graph_facts,
-                         fanout_data=fanout_graph_facts,
-                         duthost=duthost,
-                         dut_port=dut_port,
+                         fanout_data=fanout_graph_facts_multidut,
+                         dut_port=snappi_ports[0]['peer_port'],
                          prio_list=lossless_prio_list,
                          prio_dscp_map=prio_dscp_map,
-                         trigger_pfcwd=trigger_pfcwd)
-
-    config_reload(sonic_host=duthost, config_source='config_db', safe_reload=True)
+                         trigger_pfcwd=trigger_pfcwd,
+                         snappi_extra_params=snappi_extra_params)
